AskYT__Final/requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib
fpdf
googletrans==4.0.0-rc1
textblob
vaderSentiment
speechrecognition
google-cloud-speech
pyaudio